cambio de api y logica

+++ src/AbstractFactory/ClinicaFactory.java
@@ -4,6 +4,6 @@ import Modelo.Medico;
 import Modelo.Sala;
 
 public interface ClinicaFactory {
-    Medico crearMedico(int id, String nombre, String especialidad);
+    Medico crearMedico(Long id, String nombre, String especialidad);
     Sala crearSala(int numero, String descripcion);
 }
