+++ src/Observer/Observador.java
@@ -1,5 +1,5 @@
 package Observer;
 
 public interface Observador {
-    void actualizar(String nuevoEstado, int citaId);
+    void actualizar(String nuevoEstado, long citaId);
 }
